Details Activity improvements

+++ app/src/main/java/com/andre/agel/sessiontime/data/model/Movie.kt
@@ -25,7 +25,7 @@ data class Movie (
 
     val video: Boolean,
 
-    val vote_average: Double,
+    val vote_average: Float,
 
     val vote_count: Long
 )
